Initial cleanup pass, bumping things up to Java 26.

+++ vassal-app/src/main/java/VASSAL/tools/hex/Hex.java
@@ -73,7 +73,7 @@ public class Hex {
   }
 
   public int length() {
-    return (int)((Math.abs(q) + Math.abs(r) + Math.abs(s)) / 2);
+    return (Math.abs(q) + Math.abs(r) + Math.abs(s)) / 2;
   }
 
   public int distance(Hex b) {
